Consolidated 'arraytypes.compare_optional_positions()' with 'measure.within_ball()'
